# client/public/swagger.yaml
openapi: 3.0.0
info:
  description: Api documentation
  version: 1.0.1
  title: aim-container API
...
servers:
  - url: http://127.0.0.1:43801/{basePath}
    variables:
      basePath:
        default: api/v1
    description: Local env
security:
  - tokenAuth:
paths:
...
  "/executables/new":
    post:
      operationId: newExecutable
      summary: Create new executable
      tags:
        - Executable
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                script_path:
                  type: string
                arguments:
                  type: string
                env_vars:
                  type: string
                interpreter_path:
                  type: string
                working_dir:
                  type: string
                aim_experiment:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
  "/executables/kill/{pid}":
    get:
      operationId: killRunningExecutable
      summary: Kill running executables
      tags:
        - Executable
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/executables/execute":
    post:
      operationId: executeExecutable
      summary: Execute executable
      tags:
        - Executable
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                script_path:
                  type: string
                arguments:
                  type: string
                env_vars:
                  type: string
                interpreter_path:
                  type: string
                working_dir:
                  type: string
                aim_experiment:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/executables/execute/{executable_id}":
    get:
      operationId: executeExecutableTemplate
      summary: Execute executable template
      tags:
        - Executable
      parameters:
        - name: executable_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: executeExecutableTemplateForm
      summary: Execute executable from template
      tags:
        - Executable
      parameters:
        - name: executable_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                script_path:
                  type: string
                arguments:
                  type: string
                env_vars:
                  type: string
                interpreter_path:
                  type: string
                working_dir:
                  type: string
                aim_experiment:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/executables/{executable_id}":
    get:
      operationId: getExecutable
      summary: Get executable
      tags:
        - Executable
      parameters:
        - name: executable_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: saveExecutable
      summary: Save executable
      tags:
        - Executable
      parameters:
        - name: executable_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                script_path:
                  type: string
                arguments:
                  type: string
                env_vars:
                  type: string
                interpreter_path:
                  type: string
                working_dir:
                  type: string
                aim_experiment:
                  type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/executables/process/{process_id}":
    get:
      operationId: getExecutableProcess
      summary: Get executable process
      tags:
        - Executable
      parameters:
        - name: process_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/tags/{commit_id}":
    get:
      operationId: getCommitTags
      summary: Get commit tags
      tags:
        - Commit
      parameters:
          - name: commit_id
            in: path
            required: true
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/tags/update/{commit_hash}/{tag_id}":
    get:
      operationId: updateCommitTag
      summary: Update commit tag
      tags:
        - Commit
      parameters:
        - name: commit_hash
          in: path
          required: true
          schema:
            type: string
        - name: tag_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/info/{experiment}/{commit_id}":
    get:
      operationId: getCommitInfo
      summary: Get commit info
      tags:
        - Commit
      parameters:
        - name: commit_id
          in: path
          required: true
          schema:
            type: string
        - name: experiment
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
  "/commits/search?q={query}":
    get:
      operationId: getCommitsByQuery
      summary: Search commits
      tags:
        - Commit
      parameters:
          - name: query
            in: path
            required: false
            schema:
              type: string
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      name: Authorization
      in: header
...